+++ webpack.config.js
@@ -1,12 +1,12 @@
 const path = require('path');
 const webpack = require('webpack');
 const merge = require('webpack-merge');
-const WebpackNotifierPlugin = require('webpack-notifier');
 const autoprefixer = require('autoprefixer');
 const precss = require('precss');
 const flexbugs = require('postcss-flexbugs-fixes');
+const packageConfig = require('./package.json');
 
-const libraryName = 'format-utils';
+const libraryName = packageConfig.name.replace('@opuscapita/', '');
 
 const isProd = process.env.NODE_ENV === 'production';
 
@@ -43,21 +43,16 @@ const baseConfig = {
         ],
       },
       {
-        test: /\.css$/,
-        use: [
-          'style-loader',
-          'css-loader',
-          {
-            loader: 'postcss-loader',
-            options: {
-              plugins: () => [flexbugs, precss, autoprefixer],
-              minimize: !!isProd,
-            },
-          },
-        ],
+        test: /\.svg$/,
+        use: [{
+          loader: 'babel-loader',
+        },
+        {
+          loader: 'react-svg-loader',
+        }],
       },
       {
-        test: /\.scss$/,
+        test: /\.(sa|sc|c)ss$/,
         use: [
           'style-loader',
           'css-loader',
@@ -65,45 +60,51 @@ const baseConfig = {
             loader: 'postcss-loader',
             options: {
               plugins: () => [flexbugs, precss, autoprefixer],
-              minimize: !!isProd,
             },
           },
           'sass-loader',
         ],
       },
       {
-        test: /\.svg$/,
-        exclude: path.resolve(__dirname, '..', 'node_modules', 'font-awesome'),
-        use: ['babel-loader', 'react-svg-loader'],
-      },
-      {
-        test: /\.svg$/,
-        include: path.resolve(__dirname, '..', 'node_modules', 'font-awesome'),
-        use: ['file-loader?name=[name].[ext]'],
-      },
-      {
         test: /\.(woff|woff2)(\?v=\d+\.\d+\.\d+)?$/,
-        use: [
-          'url-loader?limit=100&mimetype=application/font-woff&name=fonts/[name].[ext]',
-        ],
+        use: [{
+          loader: 'file-loader',
+          options: {
+            name: '[name].[ext]',
+            outputPath: 'fonts/',
+            mimetype: 'application/font-woff',
+          },
+        }],
       },
       {
         test: /\.ttf(\?v=\d+\.\d+\.\d+)?$/,
-        use: [
-          'url-loader?limit=100&mimetype=application/octet-stream&name=fonts/[name].[ext]',
-        ],
+        use: [{
+          loader: 'file-loader',
+          options: {
+            name: '[name].[ext]',
+            outputPath: 'fonts/',
+            mimetype: 'application/octet-stream',
+          },
+        }],
       },
       {
         test: /\.eot(\?v=\d+\.\d+\.\d+)?$/,
-        use: [
-          'file-loader?name=fonts/[name].[ext]',
-        ],
+        use: [{
+          loader: 'file-loader',
+          options: {
+            name: '[name].[ext]',
+            outputPath: 'fonts/',
+          },
+        }],
       },
       {
         test: /\.ico$/,
-        use: [
-          'file-loader?name=[name].[ext]',
-        ],
+        use: [{
+          loader: 'file-loader',
+          options: {
+            name: '[name].[ext]',
+          },
+        }],
       },
     ],
   },
@@ -141,6 +142,7 @@ const baseConfig = {
 * DEVELOPMENT CONFIG
 */
 const devConfig = {
+  mode: 'development',
   devtool: 'eval-source-map',
   plugins: [
     new webpack.DefinePlugin({
@@ -148,8 +150,6 @@ const devConfig = {
         NODE_ENV: JSON.stringify('development'),
       },
     }),
-    new WebpackNotifierPlugin(),
-    new webpack.NamedModulesPlugin(),
   ],
 };
 
@@ -157,18 +157,12 @@ const devConfig = {
 * PRODUCTION CONFIG
 */
 const prodConfig = {
-  devtool: 'source-map',
+  mode: 'production',
+  devtool: false,
   plugins: [
     new webpack.DefinePlugin({
       'process.env': {
         NODE_ENV: JSON.stringify('production'),
-      },
-    }),
-    new webpack.optimize.ModuleConcatenationPlugin(),
-    new webpack.optimize.UglifyJsPlugin({
-      sourceMap: true,
-      output: {
-        comments: false,
       },
     }),
   ],